# server/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Platform for event management: authentication, user profiles, events with comments and registration",
        "title": "Eventor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://eventor.local/support",
            "email": "support@eventor.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns profile of the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get my profile",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_user_transport_http_dto_user.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update authenticated user's profile. Email and role cannot be changed via this endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update my profile",
                "parameters": [
                    {
                        "description": "Profile fields to update (all optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_user_transport_http_dto_user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_user_transport_http_dto_user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden: email/role change not allowed via this endpoint",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: email already exists",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/avatar": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload new avatar image for authenticated user. Accepts multipart/form-data.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update my avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Avatar image file (JPEG, PNG, WebP)",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar updated successfully",
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_user_transport_http_dto_user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid file or not an image",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "File storage error",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove avatar from authenticated user's profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete my avatar",
                "responses": {
                    "200": {
                        "description": "Avatar deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_user_transport_http_dto_user.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Avatar not found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{id}/avatar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns avatar of specified user as base64 data URL. Requires authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get avatar as data URL",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar data URL",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.AvatarResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User or avatar not found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/avatar/file": {
            "get": {
                "description": "Download avatar image file directly. Public endpoint (no auth required).",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get avatar as raw file",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar image file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User or avatar not found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "File read error",
                        "schema": {
                            "$ref": "#/definitions/internal_user_transport_http_controllers_user.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_user_transport_http_controllers_user.AvatarResponse": {
            "description": "Response containing avatar data (for GetAvatar endpoint)",
            "type": "object",
            "properties": {
                "data_url": {
                    "description": "Avatar as data URL (base64-encoded, for inline display)\n@example \"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...\"",
                    "type": "string"
                },
                "url": {
                    "description": "Direct URL to fetch avatar file\n@example \"/api/v1/users/12345/avatar/file\"",
                    "type": "string"
                }
            }
        },
        "internal_user_transport_http_controllers_user.ErrorResponse": {
            "description": "Standard error response format",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code\n@example 404",
                    "type": "integer"
                },
                "details": {
                    "description": "Detailed validation errors (if applicable)\n@example {\"email\": \"must be valid email\", \"name\": \"is required\"}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "description": "Human-readable error message\n@example \"user not found\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token in format: `Bearer \u003cyour_token_here\u003e`. Token obtained from /auth/login or /auth/register.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}